# Fix Resolution Bug