@rashil2000 2022-09-21
Have you checked this on multiple monitors? It fails with this error:

![image](https://user-images.githubusercontent.com/46838874/191495099-803eb428-6759-4a44-9fa9-353243d5fb68.png)