@Carterpersall 2022-09-21
> Have you checked this on multiple monitors?
> 
> You can fix it by dividing later:
>
> ```powershell
> "$($monitors[$i].Bounds.Size.Width * ($scale[$i] / 96))x$($monitors[$i].Bounds.Size.Height * ($scale[$i] / 96))"
> ```

This is how I initially had it when I was testing with multiple monitors, I changed it to the broken version later when I had only one so I didn't see this error.