@Carterpersall 2022-09-21
It seems that the way I'm currently getting the DPI always returns 120 (125%) for some reason. Guess I'm back to square one again  _sigh_