@Carterpersall 2022-09-22
It *should* be working now. Added two different implementations depending on if the script is elevated. 
The first one uses the registry location that has every saved display configuration, compares the timestamp key on each one, and takes the most recent one (current one). The location requires admin to access though which is why it checks if it is elevated beforehand. The location also includes each monitor's refresh rate, so I added that into the returned string.
The second method uses the implementation in #131 which is about 4x slower for me (50ms vs 200ms on a non-bottlenecked PowerShell instance).