@rashil2000 2022-09-22
I'm not really a fan of this method. It is extremely unlikely that anyone would run winfetch as admin (most users we get bug reports from don't even have admin rights on their PC), and even if they do, running as admin for just one function does not make much sense. Not to mention that there's essentially no way for users to know that resolution would be reported faster if they run winfetch as admin.

In essence the PR brings me back to the problems I had with #131, i.e., increased script complexity and a serious performance regression for the average use case.