@Carterpersall 2022-09-22
> It is extremely unlikely that anyone would run winfetch as admin (most users we get bug reports from don't even have admin rights on their PC)

I almost always use PowerShell as admin, but I might be an outlier there.

>running as admin for just one function does not make much sense.

I agree, but if someone is in an elevated instance already for some other reason, I don't see any reason not to use a faster method exclusive to elevated instances.

>Not to mention that there's essentially no way for users to know that resolution would be reported faster if they run winfetch as admin.

We could add a comment into the config file generator that says that a faster method is used in elevated instances
 
> In essence the PR brings me back to the problems I had with #131, i.e., increased script complexity and a serious performance regression for the average use case.

I usually would agree, but the current implementation is broken for everyone who don't have their scaling at 100% (which likely is the majority of users). And the implementation here, while slower, actually works no matter the scaling you have set in your system.