@rashil2000 2022-09-22
> We could add a comment into the config file generator that says that a faster method is used in elevated instances

Since resolution segment is enabled by default, people will not check the config file for it.


>  (which likely is the majority of users). 

It's majority of _users who use a non-default scaling_, which would actually be a small number.


Nevertheless, if we are going to merge the DLLImport code, I'd much rather prefer we do it in #131, because @HO-COOH was the original author for it.

As for this PR, we can settle on a middle ground - check for admin and use your method if true, otherwise fallback to what was present originally.