@Carterpersall 2022-09-23
> It's majority of _users who use a non-default scaling_, which would actually be a small number.

The bug occurs for everyone who has their scaling on something other than 100%. And Windows almost always has the default setting higher than 100% (for me it's 200% for some reason). This is because Windows tries to set the default scaling to match 720p, and the majority of people using this script will be running 1080p or higher. Thus, for the majority of users, the current implementation does not work correctly.

> Nevertheless, if we are going to merge the DLLImport code, I'd much rather prefer we do it in #131, because @HO-COOH was the original author for it.
> 
> As for this PR, we can settle on a middle ground - check for admin and use your method if true, otherwise fallback to what was present originally.

To prevent this pull request conflicting with the changes in #131, how about #131 gets merged first, then this pull request?