@rashil2000 2022-09-23
> To prevent this pull request conflicting with the changes in https://github.com/kiedtl/winfetch/pull/131, how about https://github.com/kiedtl/winfetch/pull/131 gets merged first, then this pull request?

Resolving conflicts should not be that difficult - so I'm fine with either