@Carterpersall 2022-09-26
I'd prefer to keep this as is and see what needs to be done once #131 is merged